+++ ItemSearchPlugin/ItemSearchWindow.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Dynamic;
 using System.Linq;
 using System.Numerics;
 using System.Runtime.InteropServices;
@@ -23,7 +24,6 @@ namespace ItemSearchPlugin {
         private readonly DalamudPluginInterface pluginInterface;
         private readonly DataManager data;
         private readonly UiBuilder builder;
-
         private Item selectedItem;
         private int selectedItemIndex = -1;
         private TextureWrap selectedItemTex;
@@ -36,8 +36,12 @@ namespace ItemSearchPlugin {
 
         public event EventHandler<Item> OnItemChosen;
         public event EventHandler<bool> OnConfigButton;
+        public event EventHandler<Item> OnMarketboardOpen;
 
         public List<ISearchFilter> searchFilters;
+
+        private bool marketBoardResponsed = false;
+
         
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
         private delegate byte TryOnDelegate(uint unknownSomethingToDoWithBeingEquipable, uint itemBaseId, byte stainColor, uint itemGlamourId, byte unknownByte);
@@ -61,6 +65,20 @@ namespace ItemSearchPlugin {
             searchFilters.Add(new LevelEquipSearchFilter(pluginConfig));
             searchFilters.Add(new LevelItemSearchFilter(pluginConfig));
             searchFilters.Add(new EquipAsSearchFilter(pluginConfig, data));
+
+            pluginInterface.Subscribe("MarketBoardPlugin", (o) => {
+                PluginLog.Log("Recieved Message from MarketBoardPlugin");
+                dynamic msg = o;
+                if (msg.Target == "ItemSearchPlugin" && msg.Action == "pong") {
+                    marketBoardResponsed = true;
+                }
+            });
+
+            dynamic areYouThereMarketBoard = new ExpandoObject();
+            areYouThereMarketBoard.Target = "MarketBoardPlugin";
+            areYouThereMarketBoard.Action = "ping";
+
+            pluginInterface.SendMessage(areYouThereMarketBoard);
 
             Task.Run(() => this.data.GetExcelSheet<Item>().GetRows()).ContinueWith(t => this.luminaItems = t.Result);
 
@@ -240,18 +258,27 @@ namespace ItemSearchPlugin {
                     Log.Error($"Exception in Choose: {ex.Message}");
                 }
             }
+            ImGui.PopStyleVar();
+
+            if (pluginConfig.MarketBoardPluginIntegration && marketBoardResponsed && this.selectedItemIndex >= 0 && this.searchTask.Result[this.selectedItemIndex].ItemSearchCategory > 0){
+                ImGui.SameLine();
+                if (ImGui.Button("Market")){
+                    OnMarketboardOpen?.Invoke(this, this.searchTask.Result[this.selectedItemIndex]);
+                }
+            }
+
 
             if (pluginConfig.SelectedDataSite != null) {
                 ImGui.SameLine();
+                ImGui.PushStyleVar(ImGuiStyleVar.Alpha, this.selectedItemIndex < 0 ? 0.25f : 1);
                 if (ImGui.Button($"View on {pluginConfig.SelectedDataSite.Name}")) {
                     if (this.selectedItemIndex >= 0) {
                         pluginConfig.SelectedDataSite.OpenItem(this.searchTask.Result[this.selectedItemIndex]);
                     }
                 }
+                ImGui.PopStyleVar();
             }
             
-            ImGui.PopStyleVar();
-
             if (!pluginConfig.CloseOnChoose) {
                 ImGui.SameLine();
                 if (ImGui.Button(Loc.Localize("Close", "Close")))
@@ -282,7 +309,7 @@ namespace ItemSearchPlugin {
         }
 
         public void Dispose() {
-
+            pluginInterface.Unsubscribe("MarketBoardPlugin");
             foreach(ISearchFilter f in searchFilters){
                 f?.Dispose();
             }